backend/requirements.txt
fastapi
uvicorn[standard]
polars
yfinance
requests
python-dotenv